edit 컬럼 타입 수정

+++ src/main/java/com/mysite/sbb/content/entity/ContentBodySchema.java
@@ -11,10 +11,10 @@ public class ContentBodySchema {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "schema_id")
-    private Long schemaId;
+    private int schemaId;
 
     @Column(name = "content_type_id")
-    private Long contentTypeId;
+    private int contentTypeId;
 
     @Column(name = "schema_type", length = 100)
     private String schemaType;
